requirements.txt
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.10
Django==1.10.3
django-cors-middleware==1.3.1
django-extensions==1.7.5
django-friendship==1.5.0
django-jet==1.0.3
djangorestframework==3.5.3
enum34==1.1.6
ipython==5.1.0
ipython-genutils==0.1.0
pathlib2==2.1.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.9
psycopg2==2.6.2
ptyprocess==0.5.1
Pygments==2.1.3
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.1
wcwidth==0.1.7
